# Refactor test.cpp