--- a/tests/flexbuffers_test.h
+++ b/tests/flexbuffers_test.h
@@ -1,0 +1,16 @@
+#ifndef TESTS_FLEXBUFFERS_TEST_H
+#define TESTS_FLEXBUFFERS_TEST_H
+
+namespace flatbuffers {
+namespace tests {
+
+void FlexBuffersTest();
+void FlexBuffersReuseBugTest();
+void FlexBuffersFloatingPointTest();
+void FlexBuffersDeprecatedTest();
+void ParseFlexbuffersFromJsonWithNullTest();
+
+}  // namespace tests
+}  // namespace flatbuffers
+
+#endif  // TESTS_FLEXBUFFERS_TEST_H