--- a/tests/fuzz_test.h
+++ b/tests/fuzz_test.h
@@ -1,0 +1,13 @@
+#ifndef TESTS_FUZZ_TEST_H
+#define TESTS_FUZZ_TEST_H
+
+namespace flatbuffers {
+namespace tests {
+
+void FuzzTest1();
+void FuzzTest2();
+
+}  // namespace tests
+}  // namespace flatbuffers
+
+#endif