Splits `tests/test.cpp` into various modules to reduce compilation time.

This of course, doesn't decrease the amount of work, but it allows parallelized builds to go much faster. Especially on incremental rebuilds.

At least on my machine, it reduced the compilation time from 13s down to 5s, for this `test.cpp` in isolation. It is actually a tad bit lower than a file I created in this PR: `monster_test.cpp` is 5.6s. 

I think the rest of this can be refactored more, but since it is no longer the biggest offending thing, I am going to devote my time to that: the `idl_parser.cpp`