@dbaileychess 2022-08-27
This is the time for the following

```
make clean
time make -j
```

On my computer, which is pretty beefy. 

With this refactor:

```________________________________________________________
Executed in   21.09 secs    fish           external
   usr time  183.36 secs  637.00 micros  183.36 secs
   sys time    6.27 secs  181.00 micros    6.27 secs
   ```

Without this refactor (at head):

```
________________________________________________________
Executed in   24.55 secs    fish           external
   usr time  168.07 secs  591.00 micros  168.07 secs
   sys time    5.50 secs  165.00 micros    5.50 secs
```

So ~3.5 seconds savings overall (~15% speedup) just for a refactor.