@aardappel 2022-08-28
It's probably a good idea to split these up.. if anything, I question the organization, either they should all be named `test_*.cpp` instead of `*_test.cpp` so they're easier to work with in the current structure, or moved to `test/cpp` etc.

Also, not sure I see the utility of having a headerfile for each, since there's nothing in them.. I'd make 1 test function per `.cpp` and just `extern` that in `test.cpp` to simplify things.