@dbaileychess 2022-08-28
The [Google style](https://google.github.io/styleguide/cppguide.html#File_Names) is to postfix `_test` to the thing being tested.

In terms of `tests/cpp` I was going to move everything Non-C++ out of `tests/` and just leave the c++ tests located here. I think I prefer the `$lang/test/` breakdown anyways.

`Extern` is frowned upon (https://google.github.io/styleguide/cppguide.html#Forward_Declarations) for this kind of thing. Yes, adding the header files seems overkill, but I think it is the correct thing to do.